+++ docs/src/demos/Experiments/MultipleEditors/index.vue
@@ -158,7 +158,7 @@ export default {
         display: flex;
         align-items: center;
 
-        > input {
+        > label {
           flex: 0 0 auto;
           margin-right: 0.5rem;
         }