+++ docs/src/demos/Nodes/TaskItem/index.vue
@@ -56,7 +56,7 @@ ul[data-type="taskList"] {
     display: flex;
     align-items: center;
 
-    > input {
+    > label {
       flex: 0 0 auto;
       margin-right: 0.5rem;
     }