+++ packages/core/src/ExtensionManager.ts
@@ -25,7 +25,7 @@ export default class ExtensionManager {
 
   constructor(extensions: Extensions, editor: Editor) {
     this.editor = editor
-    this.extensions = extensions
+    this.extensions = this.sort(extensions)
     this.schema = getSchema(this.extensions)
 
     this.extensions.forEach(extension => {
@@ -59,10 +59,6 @@ export default class ExtensionManager {
         this.editor.on('selectionUpdate', extension.config.onSelectionUpdate.bind(context))
       }
 
-      if (typeof extension.config.onViewUpdate === 'function') {
-        this.editor.on('viewUpdate', extension.config.onViewUpdate.bind(context))
-      }
-
       if (typeof extension.config.onTransaction === 'function') {
         this.editor.on('transaction', extension.config.onTransaction.bind(context))
       }
@@ -78,6 +74,22 @@ export default class ExtensionManager {
       if (typeof extension.config.onDestroy === 'function') {
         this.editor.on('destroy', extension.config.onDestroy.bind(context))
       }
+    })
+  }
+
+  private sort(extensions: Extensions) {
+    const defaultPriority = 100
+
+    return extensions.sort((a, b) => {
+      if ((a.config.priority || defaultPriority) > (b.config.priority || defaultPriority)) {
+        return -1
+      }
+
+      if ((a.config.priority || defaultPriority) < (b.config.priority || defaultPriority)) {
+        return 1
+      }
+
+      return 0
     })
   }