+++ packages/extension-collaboration-cursor/src/collaboration-cursor.ts
@@ -31,6 +31,8 @@ const awarenessStatesToArray = (states: Map<number, { [key: string]: any }>) => 
 export const CollaborationCursor = Extension.create<CollaborationCursorOptions>({
   name: 'collaborationCursor',
 
+  priority: 1000,
+
   defaultOptions: {
     provider: null,
     user: {