+++ packages/extension-paragraph/src/paragraph.ts
@@ -20,6 +20,8 @@ declare module '@tiptap/core' {
 export const Paragraph = Node.create<ParagraphOptions>({
   name: 'paragraph',
 
+  priority: 1000,
+
   defaultOptions: {
     HTMLAttributes: {},
   },