+++ packages/react/src/useReactNodeView.ts
@@ -1,12 +1,10 @@
 import { createContext, useContext } from 'react'
 
 export interface ReactNodeViewContextProps {
-  isEditable: boolean,
   onDragStart: (event: DragEvent) => void,
 }
 
 export const ReactNodeViewContext = createContext<Partial<ReactNodeViewContextProps>>({
-  isEditable: undefined,
   onDragStart: undefined,
 })