+++ packages/vue-starter-kit/src/index.ts
@@ -1,2 +1,0 @@
-export * from '@tiptap/starter-kit'
-export * from '@tiptap/vue'