+++ packages/vue/src/index.ts
@@ -1,4 +1,0 @@
-export * from '@tiptap/core'
-export { default as VueRenderer } from './VueRenderer'
-export { default as VueNodeViewRenderer } from './VueNodeViewRenderer'
-export { default as EditorContent } from './components/EditorContent'